requirements.txt
anthropic>=0.42.0
pyyaml>=6.0
